requirements.txt
pandas
numpy
scikit-learn
ta
requests
python-telegram-bot
kucoin-python
Joblib